Move card type parsing into payment method nonce inspector.
