+++ Drop-In/src/main/java/com/braintreepayments/api/DropInClient.java
@@ -334,7 +334,7 @@ public class DropInClient {
             if (configuration != null) {
                 Set<CardType> supportedCardTypes = new HashSet<>();
                 for (String cardTypeAsString : configuration.getSupportedCardTypes()) {
-                    CardType cardType = parseCardType(cardTypeAsString);
+                    CardType cardType = nonceInspector.parseCardType(cardTypeAsString);
                     if (cardType != null) {
                         supportedCardTypes.add(cardType);
                     }
@@ -459,32 +459,5 @@ public class DropInClient {
         String key = DropInResult.LAST_USED_PAYMENT_METHOD_TYPE;
         String value = nonceInspector.getPaymentMethodType(paymentMethodNonce).name();
         BraintreeSharedPreferences.getInstance().putString(context, key, value);
-    }
-
-    static CardType parseCardType(String cardType) {
-        switch (cardType) {
-            case PaymentMethodCanonicalName.AMEX:
-                return CardType.AMEX;
-            case PaymentMethodCanonicalName.DINERS_CLUB:
-                return CardType.DINERS_CLUB;
-            case PaymentMethodCanonicalName.DISCOVER:
-                return CardType.DISCOVER;
-            case PaymentMethodCanonicalName.JCB:
-                return CardType.JCB;
-            case PaymentMethodCanonicalName.MAESTRO:
-                return CardType.MAESTRO;
-            case PaymentMethodCanonicalName.MASTERCARD:
-                return CardType.MASTERCARD;
-            case PaymentMethodCanonicalName.VISA:
-                return CardType.VISA;
-            case PaymentMethodCanonicalName.UNION_PAY:
-                return CardType.UNIONPAY;
-            case PaymentMethodCanonicalName.HIPER:
-                return CardType.HIPER;
-            case PaymentMethodCanonicalName.HIPERCARD:
-                return CardType.HIPERCARD;
-            default:
-                return null;
-        }
     }
 }
